chore: extract LinkMode related code into wmbus/link_mode.cc
