+++ src/wmbus.cc
@@ -43,174 +43,6 @@
 
 using namespace std;
 
-struct LinkModeInfo
-{
-    LinkMode mode;
-    const char *name;
-    const char *lcname;
-    const char *option;
-    uint64_t val;
-};
-
-LinkModeInfo link_modes_[] = {
-#define X(name,lcname,option,val) { LinkMode::name, #name , #lcname, #option, val },
-LIST_OF_LINK_MODES
-#undef X
-};
-
-const char *toString(LinkMode lm)
-{
-#define X(name,lcname,option,val) if (lm == LinkMode::name) return #lcname;
-LIST_OF_LINK_MODES
-#undef X
-
-    return "unknown";
-}
-
-LinkModeInfo *getLinkModeInfo(LinkMode lm);
-LinkModeInfo *getLinkModeInfoFromBit(int bit);
-
-LinkModeInfo *getLinkModeInfo(LinkMode lm)
-{
-    for (auto& s : link_modes_)
-    {
-        if (s.mode == lm)
-        {
-            return &s;
-        }
-    }
-    assert(0);
-    return NULL;
-}
-
-LinkModeInfo *getLinkModeInfoFromBit(uint64_t bit)
-{
-    for (auto& s : link_modes_)
-    {
-        if (s.val == bit)
-        {
-            return &s;
-        }
-    }
-    assert(0);
-    return NULL;
-}
-
-LinkMode isLinkModeOption(const char *arg)
-{
-    for (auto& s : link_modes_) {
-        if (!strcmp(arg, s.option)) {
-            return s.mode;
-        }
-    }
-    return LinkMode::UNKNOWN;
-}
-
-LinkMode toLinkMode(const char *arg)
-{
-    for (auto& s : link_modes_) {
-        if (!strcmp(arg, s.lcname)) {
-            return s.mode;
-        }
-    }
-    return LinkMode::UNKNOWN;
-}
-
-LinkModeSet parseLinkModes(string m)
-{
-    LinkModeSet lms;
-    char buf[m.length()+1];
-    strcpy(buf, m.c_str());
-    char *saveptr {};
-    const char *tok = strtok_r(buf, ",", &saveptr);
-    while (tok != NULL)
-    {
-        LinkMode lm = toLinkMode(tok);
-        if (lm == LinkMode::UNKNOWN)
-        {
-            error(EXIT_FAILURE, "(wmbus) not a valid link mode: %s\n", tok);
-        }
-        lms.addLinkMode(lm);
-        tok = strtok_r(NULL, ",", &saveptr);
-    }
-    return lms;
-}
-
-bool isValidLinkModes(string m)
-{
-    LinkModeSet lms;
-    char buf[m.length()+1];
-    strcpy(buf, m.c_str());
-    char *saveptr {};
-    const char *tok = strtok_r(buf, ",", &saveptr);
-    while (tok != NULL)
-    {
-        LinkMode lm = toLinkMode(tok);
-        if (lm == LinkMode::UNKNOWN)
-        {
-            return false;
-        }
-        lms.addLinkMode(lm);
-        tok = strtok_r(NULL, ",", &saveptr);
-    }
-    return true;
-}
-
-LinkModeSet &LinkModeSet::addLinkMode(LinkMode lm)
-{
-    for (auto& s : link_modes_) {
-        if (s.mode == lm) {
-            set_ |= s.val;
-        }
-    }
-    return *this;
-}
-
-void LinkModeSet::unionLinkModeSet(LinkModeSet lms)
-{
-    set_ |= lms.set_;
-}
-
-void LinkModeSet::disjunctionLinkModeSet(LinkModeSet lms)
-{
-    set_ &= lms.set_;
-}
-
-bool LinkModeSet::supports(LinkModeSet lms)
-{
-    // Will return false, if lms is UKNOWN (=0).
-    return (set_ & lms.set_) != 0;
-}
-
-bool LinkModeSet::has(LinkMode lm)
-{
-    LinkModeInfo *lmi = getLinkModeInfo(lm);
-    return (set_ & lmi->val) != 0;
-}
-
-bool LinkModeSet::hasAll(LinkModeSet lms)
-{
-    return (set_ & lms.set_) == lms.set_;
-}
-
-string LinkModeSet::hr()
-{
-    string r;
-    if (set_ == Any_bit) return "any";
-    if (set_ == 0) return "none";
-    for (auto& s : link_modes_)
-    {
-        if (s.mode == LinkMode::Any) continue;
-        if (set_ & s.val)
-        {
-            r += s.lcname;
-            r += ",";
-        }
-    }
-    r.pop_back();
-    return r;
-}
-
 struct Manufacturer {
     const char *code;
     int m_field;
@@ -4176,17 +4008,6 @@ uint64_t dataAsUint64(int dif, int vif, int vife, string data)
     case 0xF: return -1; // Special Functions
     }
     return -1;
-}
-
-string linkModeName(LinkMode link_mode)
-{
-
-    for (auto& s : link_modes_) {
-        if (link_mode == s.mode) {
-            return s.name;
-        }
-    }
-    return "UnknownLinkMode";
 }
 
 string measurementTypeName(MeasurementType mt)
